\name{dir_}
\alias{libinv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
libinv
}
\description{
Loads list of packages in silent mode
}
\usage{
libinv(packages)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{packages}{
Vector of packages names (strings)
}
}
\details{
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Valerio Barbarossa
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
libinv(packages)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~packages }% use one of  RShowDoc("KEYWORDS")
