\name{rm_}
\alias{rm_}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rm_
}
\description{
Removes a directory recursively
}
\usage{
rm_(name_dir)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{name_dir}{
Path to directory
}
}
\details{
For Windows and Linux (or wherever "rm -r" works)
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Valerio Barbarossa
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
rm_(name_dir)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~directory }% use one of  RShowDoc("KEYWORDS")
